requirements.txt
Django>=3.2.18,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15,<0.16
Pillow>=8.2.0,<8.3.0